Remove back the gain of sqrt(L) from the interpolator LPF
The gain is only valid when we are dealing with DTFT. Since htx is designed by sampling the IFT (rather than using IDTFT), it already provides the same effect. In CT, the gain does not have to be adjusted by sqrt(L), because the frequency axis is absolute, not relative, so the bandwidth continues to be pi/Tsym after upsampling and the gain of sqrt(Tsym) already yields unitary energy. In the DTFT, however, after upsampling the normalized bandwidth is reduced by L, so sqrt(L) restores unitary energy.