Reduce the TEQ optimization search and the length/delay choice
Previously the range of lengths analyzed by the optimization fcn
was very large for the MMSE-LE TEQ. The range was significantly
reduced by this update. Now, it is restricted to a length less
than twice the CP length.

The other modification concerns the choice of the length/delay.
Previously the fcn returned the length/delay corresponding to the
maximum of the objective function. Now, it returns the minimum
length/delay among all points that are within 5% of the maximum.
It should be sufficient and prevent the highest length to be
always returned.