Change modulation order argument of the moduleOperator to b_bar
Previously the function had to infer the number of dimensions in
each of the subchannels and use sqrt(M) or M accordingly, for two
dimensional subchannels and one-dimensionals subchannels,
respectively. Instead of doing so, it is better to pass the vector
of bits/dim directly, so the function does not have to infer the
dimensions.