Adapt precoders for the shift by tau in Hisi and no shift in Hcirc
The formulation in the paper was updated and the Hcirc matrix is
now already constructed with h_n0 as the first sample, namely Hcirc
is already circularly shifted. Hence, the additional phase shift
becomes unnecessary. Also, the ISI is now stated in terms of a
rotation by tau. The Precoders were adapted accordingly.