Review the mechanism for preventing ICPD precoding power increase
Previously the pre-compensation to avoid the power increase
incurred in the ICPD precoding process was based on the individual
precoder row norms for each subchannel. Since this is only valid
for the case of flat energy load, it needed to be updated. The new
version computes the power increase in each subchannel given an
initial energy load.